history, cleanup, import

+++ bin/sysclean.sh
@@ -1,3 +1,6 @@
 #!/usr/bin/env bash
 
-find ~/ -type f \( -name ".DS_Store" -o -name ".localized" \) -exec rm -f {} \;
+if [ "$(uname)" == "Darwin" ]; then
+  find ~/ -type f \( -name ".DS_Store" -o -name ".localized" \) -exec rm -f {} \;
+  rm -rf ~/Library/Logs/*
+fi
